requirements.txt
fastapi
uvicorn[standard]
pillow
numpy<2.0
python-multipart
python-dotenv
ultralytics
easyocr
opencv-python-headless
